Hito 3

+++ Game/Game/src/components/General Components/Animator.h
@@ -1,10 +1,7 @@
 #pragma once
-
 #include "Image.h"
 #include <unordered_map>
 #include <unordered_set>
-
-
 using AnimationMap = unordered_map<string, Animation>;
 
 class Animator : public Image {
@@ -29,7 +26,7 @@ public:
 	// Crea una animacion nueva
 	void createAnim(string key, Animation anim);
 	// Crea una animacion nueva
-	void createAnim(string key, int start, int end, int rate, int _rep = 0);
+	void createAnim(string key, int start, int end, float rate, int _rep = 0);
 
 	// Empieza una nueva animacion
 	void play(string key);
